Move camera to the domain layer
