+++ domain/camera/.gitignore
