+++ domain/camera/build.gradle
