+++ domain/camera/proguard-rules.pro
