+++ domain/camera/src/androidTest/java/com/google/jetpackcamera/domain/camera/ExampleInstrumentedTest.kt
@@ -14,20 +14,27 @@
  * limitations under the License.
  */
 
-package com.google.jetpackcamera.data.camera
+package com.google.jetpackcamera.domain.camera
+
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.ext.junit.runners.AndroidJUnit4
 
 import org.junit.Test
+import org.junit.runner.RunWith
 
 import org.junit.Assert.*
 
 /**
- * Example local unit test, which will execute on the development machine (host).
+ * Instrumented test, which will execute on an Android device.
  *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
-class ExampleUnitTest {
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
     @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("com.google.jetpackcamera.data.camera.test", appContext.packageName)
     }
 }
