+++ domain/camera/src/main/AndroidManifest.xml
