+++ domain/camera/src/main/java/com/google/jetpackcamera/domain/camera/CameraRepository.kt
@@ -14,22 +14,28 @@
  * limitations under the License.
  */
 
-package com.google.jetpackcamera.data.camera
+package com.google.jetpackcamera.domain.camera
 
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
+import androidx.camera.core.CameraSelector
+import androidx.camera.core.Preview
+import androidx.lifecycle.LifecycleOwner
 
 /**
- * Dagger [Module] for camera data layer.
+ * Data layer for camera.
  */
-@Module
-@InstallIn(SingletonComponent::class)
-interface CameraModule {
+interface CameraRepository {
 
-    @Binds
-    fun bindsCameraRepository(
-        cameraXCameraRepository: CameraXCameraRepository
-    ): CameraRepository
+    /**
+     * Initializes the camera.
+     */
+    suspend fun initialize(): List<Int>
+
+    /**
+     * Starts the camera preview.
+     */
+    fun startPreview(
+        lifecycleOwner: LifecycleOwner,
+        surfaceProvider: Preview.SurfaceProvider,
+        @CameraSelector.LensFacing lensFacing: Int
+    ): Boolean
 }
