+++ domain/camera/src/test/java/com/google/jetpackcamera/domain/camera/ExampleUnitTest.kt
@@ -14,27 +14,20 @@
  * limitations under the License.
  */
 
-package com.google.jetpackcamera.data.camera
-
-import androidx.test.platform.app.InstrumentationRegistry
-import androidx.test.ext.junit.runners.AndroidJUnit4
+package com.google.jetpackcamera.domain.camera
 
 import org.junit.Test
-import org.junit.runner.RunWith
 
 import org.junit.Assert.*
 
 /**
- * Instrumented test, which will execute on an Android device.
+ * Example local unit test, which will execute on the development machine (host).
  *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
+class ExampleUnitTest {
     @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.google.jetpackcamera.data.camera.test", appContext.packageName)
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
     }
 }
